third_party/WebKit/LayoutTests/fast/workers/worker-nested-importScripts-error-expected.txt

This tests that errors from nested importScripts have the expected provenance.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/importScripts-1.js
PASS event.type is "error"
PASS event.filename.indexOf('invalidScript.js') >= 0 is true
PASS successfullyParsed is true

TEST COMPLETE
